move uploader into seedclient
