+++ .github/workflows/ci.yml
@@ -70,8 +70,7 @@ jobs:
         run: |
           # break up into the individual files because we are getting race conditions
           # these are the "integration" marked tests
-          pytest tests/test_seed_properties.py -s
-          # pytest tests/test_seed_uploader.py -s
+          pytest tests/test_seed_client.py -s
           # pytest tests/test_seed_base.py -s
       - name: Dump docker logs on failure
         if: failure()
