+++ README.rst
@@ -7,11 +7,46 @@ Py-SEED
 .. image:: https://badge.fury.io/py/py-seed.svg
     :target: https://pypi.python.org/pypi/py-seed/
 
-A python API client for the SEED Platform
-
+A python API client for the SEED Platform. This is an updated version of the Client. It is compatible with the latest version of the SEED Platform (>2.14.0). This client still has access to the previous format of generating a lower level API client by accessing `seed_client_base.SEEDOAuthReadOnlyClient`, `seed_client_base.SEEDOAuthReadWriteClient`, `seed_client_base.SEEDReadOnlyClient`, and `seed_client_base.SEEDReadWriteClient`. This lower level API is documented below under the `Low-Level Documentation`
 
 Documentation
 -------------
+The SEED client is a read-write client. To install the client run:
+
+.. code-block:: bash
+
+    pip install pyseed
+
+Within Python you can use the client like this:
+
+.. code-block:: python
+
+    import pyseed
+
+    # The seed-config.json file defines the hosting locaiton and credentials for your SEED instance.
+    # If running SEED locally for testing, then you can run the following from your SEED root directory:
+    #    ./manage.py create_test_user_json --username user@seed-platform.org --file ./seed-config.json --pyseed
+
+    config_file = Path('seed-config.json')
+    seed_client = SeedProperties(cls.organization_id, connection_config_filepath=config_file)
+
+    # Get/create the new cycle and upload the data. Make sure to set the cycle ID so that the
+    # data end up in the correct cycle
+    cycle = seed_client.get_or_create_cycle(
+        'pyseed-api-test', date(2021, 6, 1), date(2022, 6, 1), set_cycle_id=True
+    )
+
+    seed_client.upload_and_match_datafile(
+        'pyseed-properties-test',
+        'tests/data/test-seed-data.xlsx',
+        'Single Step Column Mappings',
+        'tests/data/test-seed-data-mappings.csv'
+    )
+
+    # See the projects unit tests for more examples.
+
+Low-Level Documentation
+-----------------------
 This provides two user authentication based Python clients and two OAuth2 authentication based Python clients for interacting with the SEED Platform Api::
 
 
@@ -19,6 +54,7 @@ This provides two user authentication based Python clients and two OAuth2 authen
     SEEDOAuthReadWriteClient
     SEEDReadOnlyClient
     SEEDReadWriteClient
+
 
 
 (The OAuthMixin is constructed around the the JWTGrantClient found in jwt-oauth2lib. see https://github.com/GreenBuildingRegistry/jwt_oauth2)
@@ -58,8 +94,6 @@ Usage:
     seed_client.get(property_pk, endpoint='properties')
 
 
-Contributing
-------------
 
 License
 -------
@@ -67,7 +101,6 @@ py-SEED is released under the terms of the MIT license. Full details in LICENSE 
 
 Changelog
 ---------
-py-SEED was developed for use in the greenbuildingregistry project.
-For a full changelog see `CHANGELOG.rst <https://github.com/GreenBuildingRegistry/py-seed/blob/master/CHANGELOG.rst>`_.
+py-SEED was developed for use in the greenbuildingregistry project but has been extended for various uses, including Salesforce data transfer and SEED data analysis.
 
-N.B. this client is undergoing development and should be considered experimental.
+For a full changelog see `CHANGELOG.rst <https://github.com/seed-platform/py-seed/blob/master/CHANGELOG.rst>`_.
